requirements.txt
attrs==17.2.0
Automat==0.6.0
constantly==15.1.0
hyperlink==17.3.0
incremental==17.5.0
MySQL-python==1.2.5
PyYAML==3.12
six==1.10.0
Twisted==17.5.0
zope.interface==4.4.2
